requirements.txt
alabaster==0.7.12
anytree==2.8.0
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==20.1.0
asn1crypto==1.4.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
Beaker==1.11.0
bleach==3.2.1
btrfsutil==5.11.1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
dnspython==1.16.0
docutils==0.16
entrypoints==0.3
escrotum==1.0.1
gobject==0.1.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gspread==3.7.0
html5lib==1.1
httplib2==0.19.1
idna==3.1
imagesize==1.2.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
louis==3.17.0
lxml==4.6.3
Mako==1.1.4
mallard-ducktype==1.0.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mpmath==1.1.0
msgpack==1.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
ordered-set==4.0.2
packaging==20.9
pandocfilters==1.4.3
parso==0.7.1
pep517==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
ply==3.11
progress==1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.15.8
ptyprocess==0.6.0
pwquality==1.4.4
py-trello==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycparser==2.20
pydantic==1.8.1
Pygments==2.8.1
PyGObject==3.40.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytrello==0.1.6
pytz==2021.1
pyzmq==20.0.0
ranger-fm==1.9.3
Reflector==2021.4.3.10.43.28
requests==2.25.1
requests-oauthlib==1.3.0
resolvelib==0.5.4
retrying==1.3.3
rsa==4.7.2
Send2Trash==1.5.0
sip==4.19.25
six==1.15.0
snowballstemmer==2.1.0
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sympy==1.7
team==1.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
xcffib==0.11.1
youtube-dl==2021.4.7
